+++ iocore/net/quic/QUICPathValidator.h
@@ -41,7 +41,7 @@ public:
 
   // QUICFrameGeneratro
   bool will_generate_frame() override;
-  QUICFrameUPtr generate_frame(uint16_t connection_credit, uint16_t maximum_frame_size) override;
+  QUICFrameUPtr generate_frame(uint64_t connection_credit, uint16_t maximum_frame_size) override;
 
 private:
   enum class ValidationState : int {
